Added bootstrap

+++ EventEasePOE2.0/Controllers/BookingsController.cs
@@ -88,9 +88,14 @@ namespace EventEasePOE2._0.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var booking = await _context.Bookings.FindAsync(id);
+            if (booking == null)
+            {
+                return NotFound();
+            }
             _context.Bookings.Remove(booking);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
+
     }
 }
